Making the engine not have a precheck for the round not started
